move hexColourCalculator.js file
