+++ src/client/components/SupplierProfileStrength/hexColourCalculator.js
